resize, gaps: support percentages of the monitor size
'resize width 50%' makes the window occupy half of the monitor, and
'resize width +10%' grows it by 10% of the monitor size. The same PixelsOrPercent
value that 'accordion-padding' already accepts is now also accepted by the 'gaps'
settings: horizontal gaps are resolved against the monitor width, and vertical
gaps against the monitor height.

https://github.com/nikitabobko/AeroSpace/issues/397

+++ Sources/AppBundle/model/MonitorEx.swift
@@ -4,10 +4,10 @@ extension MonitorInfo {
         let topLeft = visibleRect.topLeftCorner
         let gaps = ResolvedGaps(gaps: config.gaps, monitor: self)
         return Rect(
-            topLeftX: topLeft.x + gaps.outer.left.toDouble(),
-            topLeftY: topLeft.y + gaps.outer.top.toDouble(),
-            width: visibleRect.width - gaps.outer.left.toDouble() - gaps.outer.right.toDouble(),
-            height: visibleRect.height - gaps.outer.top.toDouble() - gaps.outer.bottom.toDouble(),
+            topLeftX: topLeft.x + gaps.outer.left,
+            topLeftY: topLeft.y + gaps.outer.top,
+            width: visibleRect.width - gaps.outer.left - gaps.outer.right,
+            height: visibleRect.height - gaps.outer.top - gaps.outer.bottom,
         )
     }
